# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sovrane API",
    "version": "1.0.0",
    "description": "API documentation for the Sovrane application"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/config/{key}": {
      "get": {
        "summary": "Get configuration value",
        "description": "Return the value for the specified configuration key. Requires the `read-config` permission.",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifier of the configuration entry."
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigEntry"
                }
              }
            }
          },
          "404": {
            "description": "Configuration entry not found"
          }
        }
      },
      "put": {
        "summary": "Update configuration value",
        "description": "Update the value of a configuration entry. Requires the `update-config` permission.",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "New value and user identifier performing the update.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "description": "New value to store."
                  },
                  "updatedBy": {
                    "type": "string",
                    "description": "Identifier of the user updating the value."
                  }
                },
                "required": [
                  "value",
                  "updatedBy"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Configuration updated successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "delete": {
        "summary": "Delete configuration value",
        "description": "Remove a configuration entry. Requires the `delete-config` permission.",
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifier of the configuration entry."
          }
        ],
        "requestBody": {
          "description": "Identifier of the user performing the deletion.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deletedBy": {
                    "type": "string",
                    "description": "Identifier of the user deleting the value."
                  }
                },
                "required": [
                  "deletedBy"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Configuration deleted successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ConfigEntry": {
        "description": "Application configuration entry.",
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Configuration key."
          },
          "value": {
            "description": "Stored configuration value.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "key",
          "value"
        ]
      },
...
    }
  },
...
}